Created Final ERD File and Commit Note
Though it requires looking over to ensure it is correct, the final ERD has been started and is, at least, close to completion.
Also, please note that commits from updating draw.io files will not have descriptions directly - for this, apologies are in order. Draw.io forces commits when saving to a GitHub folder, but includes no way of adding a description.