Updated Use Case Diagram User
It has been decided that the database for this project is being designed for a law enforcement investigator looking for threatening messages in social media - they have been given the (coursework's) dataset to look through for a case. Because of this decision, the label "User" has been changed to "Social Media Investigator".