Created Activity Diagram
The activity diagram has been created. There are two connections to the exit, both involving the finding of something suspicious by the SMI (or the finding of nothing at all). Four activities and two decisions are contained within the diagram, and the accessing of friends loops back to viewing a user (as friends are themselves users).