Amended User Stories and Normalisation
Now that it has been decided who the database is for, the user story document now has a user story (that being of the SMI, of course).
Having been on MySQL Workbench, about to create the database for the project, it has been realised that there was no relation between users and friends in the normalisation except for through messages, which shouldn't be the sole connection between them. It was initially decided to remove the "Sender ID" and "Receiver ID" as attributes as they could be replaced with the already-exisiting "User ID" and "Friend ID". However, this has been reverted as including them allows the normalisation to include a table that contains a direct link between users and friends - this is (arguably) beneficial for the database.