Updated Info Document
The info document now contains further tables and updated previous tables. For those tables that were updated (and for which it was appropriate) there is now the inclusion of coding in foreign keys.
NOTE: the activity diagram may need to be changed - it appears that decisions must actually contain conditions; it would seem that the diagram that has been produced has one too many bubbles for each decision diamond (where the content of that bubble would actually belong inside the diamond). This will have to be checked.