Updated Use Cases (Added a User)
According to recent information, it appears the diagrams must not only reflect the functions of the database, but also the four functions (Create, Read, Update, and Delete) abbreviated to CRUD, despite their absence in the database itself. This therefore involves the addition of another user - they have been called the Third-Party Database Manager. This person has been imagined as someone who owns, or at least manages, the database and has thought, for one reason or another, that within the sample section of the database info there lies suspicious conversation. Therefore, they both have access to CRUD functions and have a direct connection to the SMI.