Updated Sequence Diagram (Added TPDM)
The file has been updated to include the TPDM, meaning that the first diagram is, as such, repeated due to the TPDM having the same capabilities as the SMI and then some. Although it is not found to be completely convincing with the three new methods (createUser, updateUser, and deleteUser) simply returning to the beginning, it is unsure at present exactly how else to present them.