Updated Activity Diagram(s)
A second diagram has been added for the TPDM. It is admittedly a tad more complexed than intended, and it may need to be revised at a later date (especially with the crossing lines). The majority of the diagram is decisions, with only one activity (rounded rectangle) present. It involves both the reading done by the SMI (as so too can the TPDM do it) and the CRUD functions only the TPDM can do.