Altered Sample Code
The sample C# Database Connection Code has been edited to fit with this coursework's database. Furthermore, three more Data View Grids have been added in order to allow one for each table in the database. Ideally, selecting a user in one table will show only that user's data in the rest of them; however, enabling this to be the case could be a challenge.