Introduced Shortened Data
The messages, universities, and workplace Excel files have been edited and saved as separate files that contain only the first five-thousand users. This admittedly took a while due to personal ineptitude. Partially, this was because of MySQL's counting system being different to 'regular' counting which was initially irritating. Realising after a while that it wasn't necessary to attempt to change this, as the data itself wouldn't affect anything in any negative manner, the files were settled on (from basically creating the same ones three times in a row).
Please note also that yesterday and today have also been spent importing data (and indeed, it has been noted that it is a tad late to be doing so).
In recent silences, it should also be noted, the database itself was being worked on, though it isn't saved on this GitHub (as it wasn't known how to). Details will likely be found in the report - all SQL code to create tables and keys was personally completed (not using the wizard).