Added Methods to Class Diagram and Almost Completed Report
The report is completely finished bar having the GitHub repository itself attached to it. The README requires updating and this is to be done through the origin (hence why this commit is being done). When the README is complete, a link to the GitHub shall be provided on the report and it shall be complete.
Also note that the methods from the sequence diagram for updating, creating, and deleting had been added to the class diagram before finishing the report.